Cleanup and fix LRUCache implementation and tests
Some modernization refactoring, documentation, and some bugfixing as revealed by the unit tests.

Removed the bool that Mike had added to LRUCache in favor of just checking the null status of the objectSizes object, like is done elsewhere in the code.

Had to fix the size accumulator not getting size amounts when objectSizes wasn't active: the rest of the code assumed it would be incremented by a size of 1 and it wasn't.