config: Converted debug-mode warning for bad async calls to compiler errors
Put this this in Debug not Release so that Release code can proceed as it has always done.

This tells the compiler that any time you have a a call to an async method that doesn't await the call to treat it as a compile time error.  The problem with async calls that are not awaited? if they have an exception, that exception vaporizes into the void and is not reported anywhere..