fix: Detective work and conversion to a new explicit FireAndForget calls
A good portion of this code doesn't need to wait. It just needs to log the uncaught nested exceptions so they can be detected and fixed.

Many places the ability for an async call to be used as a threadless fire-and-forget was being used, but the annotations for that concept `Task task = asyncMethod()` without using the variable were lost in a cleanup by Cinder quite some time ago.  Now I'm introducing a cleaner, better markup that makes it explicit that this async is a fire-and-forget call.