cd: Switch to GH Actions from AppVeyor
Typescript is your friend - even though it can be a bear to get set up, and I might have been able to solve these problems with C#...

I started out in the draft creator using the GraphQL interface to pull the list of releases, however I found that while it executed, it never actually had the releases. This resulted in the code creating excess drafts. Dropping to pure REST simplifies the code and corrects the problem, at the expense of getting a LOT of data over the wire. REST just sends a LOT more info than I need.